requirements.txt
fastapi==0.119.0
uvicorn==0.37.0
pydantic==2.12.2
httpx==0.27.2
requests==2.32.3
redis==5.2.1
qdrant-client==1.11.3
loguru==0.7.2
python-dotenv==1.0.1
email-validator==2.2.0
# Agentic stack
langchain==0.3.27
langgraph==0.2.60
openai==1.109.1
transformers==4.57.1
# Comms
twilio==9.3.4
sendgrid==6.11.0
google-api-python-client==2.149.0
google-auth-oauthlib==1.2.1
# Web3
web3==6.20.2
# IPFS CLI integration relies on system ipfs; no py dep needed

# Tests
pytest==8.3.3
pytest-asyncio==0.24.0
